feat: allow to customize table column widths
