+++ packages/golden_toolkit/lib/golden_toolkit.dart
@@ -11,7 +11,7 @@ export 'src/configuration.dart';
 export 'src/device.dart';
 export 'src/device_builder.dart';
 export 'src/font_loader.dart' show loadAppFonts;
-export 'src/golden_builder.dart';
+export 'src/golden_builder/golden_builder.dart';
 export 'src/multi_screen_golden.dart';
 export 'src/testing_tools.dart' hide compareWithGolden;
 export 'src/widget_tester_extensions.dart';
